requirements.txt
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
cssselect==1.1.0
Django==2.2.4
django-haystack==3.0
django-widget-tweaks==1.4.8
DjangoUeditor==1.8.143
docxtpl==0.11.2
idna==2.10
jieba==0.42.1
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
nose==1.3.7
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pyquery==1.4.1
python-docx==0.8.10
pytz==2020.1
qrcode==6.1
requests==2.25.0
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
Whoosh==2.7.4
